requirements.txt
twisted
findtools
ipython
zmq
pytaglib
gevent
